Final bit of compiling on Windows
