+++ containerd/main_windows.go
@@ -1,0 +1,10 @@
+package main
+
+// TODO Windows: May be able to factor out entirely
+func checkLimits() error {
+	return nil
+}
+
+// No idea how to implement this on Windows.
+func reapProcesses() {
+}
